Add extensible support for handle systems and URIs
